removed the client folder
