+++ server/db/models/mentee.js
@@ -3,25 +3,58 @@
 const {
   Model
 } = require('sequelize');
+const bcrypt = require('bcryptjs');
 
 module.exports = (sequelize, DataTypes) => {
   class Mentee extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
+    validatePassword(password) {
+      return bcrypt.compareSync(password, this.hashedPassword.toString());
+    }
+
+    static getCurrentMentorById(id) {
+      return Mentee.scope('currentMentee').findByPk(id);
+    }
+
+    static async login({ email, password }) {
+      const mentee = await Mentee.scope('loginMentee').findOne({
+        where: {
+          email
+        },
+      });
+
+      if (mentee && mentee.validatePassword(password)) {
+        return await Mentee.scope('currentMentee').findByPk(mentee.id);
+      }
+    }
+
+    static async signup({
+      name, email, password, profileImg, role, about
+
+    }) {
+      const hashedPassword = bcrypt.hashSync(password);
+      const mentee = await Mentee.create({
+        name,
+        email,
+        hashedPassword,
+        profileImg,
+        role,
+        about,
+
+      });
+      return await Mentee.scope('currentMentee').findByPk(mentee.id);
+    }
+
     static associate(models) {
       Mentee.belongsTo(models.Mentor, {
         foreignKey: 'mentorId',
       });
-      Mentee.hasMany(models.Message, {
-        foreignKey: 'ownerMen',
-        onDelete: 'CASCADE',
-      });
-      Mentee.hasMany(models.Channel, {
-        foreignKey: 'menteeId',
-      });
+      // Mentee.hasMany(models.Message, {
+      //   foreignKey: 'ownerMen',
+      //   onDelete: 'CASCADE',
+      // });
+      // Mentee.hasMany(models.Channel, {
+      //   foreignKey: 'menteeId',
+      // });
       Mentee.hasMany(models.Task, {
         foreignKey: 'menteeId',
       });
@@ -32,7 +65,12 @@ module.exports = (sequelize, DataTypes) => {
       type: DataTypes.STRING,
       allowNull: false,
       validate: {
-        len: [2, 40],
+        len: [3, 40],
+        // isNotEmail(value) {
+        //   if (Validator.isEmail(value)) {
+        //     throw new Error('Cannot be an email.');
+        //   }
+        // }
       },
     },
     email: {
@@ -50,6 +88,10 @@ module.exports = (sequelize, DataTypes) => {
         len: [60, 60],
       },
     },
+    classification: {
+      type: DataTypes.STRING,
+      defaultValue: 'Mentee'
+    },
     city: DataTypes.STRING,
     state: DataTypes.STRING,
     country: DataTypes.STRING,
@@ -66,6 +108,21 @@ module.exports = (sequelize, DataTypes) => {
   }, {
     sequelize,
     modelName: 'Mentee',
+    defaultScope: {
+      attributes: {
+        exclude: ['hashedPassword', 'createdAt', 'updatedAt'],
+      },
+    },
+    scopes: {
+      currentMentee: {
+        attributes: {
+          exclude: ['hashedPassword'],
+        },
+      },
+      loginMentee: {
+        attributes: {},
+      },
+    },
   });
   return Mentee;
 };
