+++ server/db/seeders/20230509213851-demo-mentee.js
@@ -13,8 +13,9 @@ module.exports = {
       name: 'John Doe',
       email: 'john.doe@cgu.io',
       hashedPassword: bcrypt.hashSync('password'),
+      classification: 'Mentee',
       city: 'New York',
-      state: 'NY',
+      state: 'New York',
       country: 'United States',
       profileImg: 'https://images.unsplash.com/photo-1592194996308-7b43878e84a6?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxzZWFyY2h8OXx8Y2F0fGVufDB8fDB8fA%3D%3D&auto=format&fit=crop&w=500&q=60',
       goal: 'Backend Engineer',
@@ -27,6 +28,7 @@ module.exports = {
       name: 'Dag Denness',
       email: 'ddenness6@jigsy.com',
       hashedPassword: bcrypt.hashSync('password1'),
+      classification: 'Mentee',
       city: 'Whitehorse',
       state: 'Yukon Territory',
       country: 'Canada',
@@ -41,6 +43,7 @@ module.exports = {
       name: 'Glori cornhill',
       email: 'gcornhill3@artisteer.com',
       hashedPassword: bcrypt.hashSync('password2'),
+      classification: 'Mentee',
       city: 'Madison',
       state: 'Wisconsin',
       country: 'United States',
@@ -55,6 +58,7 @@ module.exports = {
       name: 'Mozelle Scandroot',
       email: 'mscandrootc@abc.net.au',
       hashedPassword: bcrypt.hashSync('password3'),
+      classification: 'Mentee',
       city: 'North Little Rock',
       state: 'Arkansas',
       country: 'United States',
